requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-gitlab==3.15.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
httpx==0.25.2
aiofiles==23.2.1
click==8.1.7

# Phase 2: Docker and sandbox execution
docker>=7.0.0
websockets>=12.0
